Use relative URLs everywhere
Since we renamed the repository from fsi-tue.github.io to cyberauftritt
the relative URLs didn't work properly anymore. The reason for this
renaming was, that GitHub Pages didn't work on a fork if the name of the
repository was "fsi-tue.github.io" (most likely the dots where
problematic).

This solution isn't ideal either since forks will have the URL from the
upstream repository as baseurl but at this can be changed manually
(and without much effort).

In the end we should try to get rid of the following two lines:
url: "https://fsi-tue.github.io"
baseurl: "/cyberauftritt"