Unset the "url" and "baseurl" configuration options
This should not break the site and will fix most of the failing
html-proofer checks. We also don't actually need or want want this crap
in the first place, especially "url" as that makes it difficult to test
changes with a local server or on forks (as most links would redirect
the user to the "real" site, i.e.
https://fsi-tue.github.io/cyberauftritt/).

TODO: Replace all occurrences of "site.url" and "site.baseurl" with
relative links.