Set the page.id variable for 404.html
I've used 100000 to avoid collisions. This is a bit stupid but it will
fix the html-proofer checks. The problem was that the code in
"_includes/head.html" will drop the ".html" extension if the variable
"page.id" is not set (which is normally ok, but we need
"permalink: /404.html" [0] which will generate a file "_site/404.html"
instead of "_site/404/index.html"):

```
  {% if page.id %}
  <meta property="og:url" content="{{ site.url }}{{ page.url }}" />
  <link rel="canonical" href="{{ site.url }}{{ page.url }}" />
  {% else %}
  <meta property="og:url" content="{{ site.url }}{{ page.url | remove: '/index.html' | remove: '.html' }}" />
  <link rel="canonical" href="{{ site.url }}{{ page.url | remove: '/index.html' | remove: '.html' }}" />
  {% endif %}
```

[0]: https://help.github.com/articles/creating-a-custom-404-page-for-your-github-pages-site/