CI: Replace Travis CI with GitHub Actions
Switching to GitHub Actions will give us more flexibility/features,
integrates better with GitHub, and should even be a faster (due to the
faster queuing).

This workflow is based on GitHub's Jekyll template [0] but with
additional fixes/improvements/simplifications, an html-proofer check,
and actions/upload-artifact@v2 for uploading the result (_site/).

For compatibility with our Dockerfile I've switched from
jekyll/builder:latest (Jekyll 4.x) to jekyll/jekyll:3.

Note: Gemfile.lock shouldn't have to be world-writable but the
unofficial jekyll/builder image [1] currently requires it. To improve
this situation it would be better to call "bundle install" [2] with the
"--deployment" flag [3]. But as long as Gemfile.lock is up-to-date with
Gemfile changes it should still work without a world-writable
Gemfile.lock (I need to check if our Gemfile.lock is up-to-date).

TODOs:
- Replace the Travis CI links
- Optional: Try to accelerate CI (e.g. via caching).

[0]: https://github.com/actions/starter-workflows/blob/8beb802437927d71bbb91605d15491672edf222a/ci/jekyll.yml
[1]: https://github.com/envygeeks/jekyll-docker
[2]: https://github.com/envygeeks/jekyll-docker/blob/575057ab5f14055e161b871e75f8295afd571aa9/repos/jekyll/copy/all/usr/jekyll/bin/jekyll#L27
[3]: https://bundler.io/man/bundle-install.1.html#DEPLOYMENT-MODE